Remove slashes from mkdir and cd commands
In PowerShell on Windows, this created a new directory at the top level,
e.g. C:\eleventy-project-name\. In Zsh on Ubuntu, the trailing slash
made a new line in the terminal and waited for me to press <enter>
again. I don't think there should be an issue without the slashes.